Change strategy.
