+++ coil-core/src/commonMain/kotlin/coil3/intercept/EngineInterceptor.kt
@@ -24,11 +24,13 @@ import coil3.request.SuccessResult
 import coil3.request.allowConversionToBitmap
 import coil3.request.transformations
 import coil3.transform.Transformation
+import coil3.util.DelayedDispatchCoroutineDispatcher
 import coil3.util.ErrorResult
 import coil3.util.Logger
 import coil3.util.SystemCallbacks
 import coil3.util.addFirst
 import coil3.util.closeQuietly
+import coil3.util.dispatcher
 import coil3.util.eventListener
 import coil3.util.foldIndices
 import coil3.util.isPlaceholderCached
@@ -69,6 +71,12 @@ internal class EngineInterceptor(
             // Fast path: return the value from the memory cache.
             if (cacheValue != null) {
                 return memoryCacheService.newResult(chain, request, cacheKey, cacheValue)
+            }
+
+            // Re-enable dispatching before starting to fetch.
+            val dispatcher = coroutineContext.dispatcher
+            if (dispatcher is DelayedDispatchCoroutineDispatcher) {
+                dispatcher.dispatchEnabled = true
             }
 
             // Slow path: fetch, decode, transform, and cache the image.
